requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.1
pydantic-settings==2.6.0
strawberry-graphql==0.246.1
aiosqlite==0.20.0
httpx==0.27.2
pytest==8.3.2